yx_recommender_system/requirements.txt
scikit-surprise==1.1.0
numpy==1.18.1
pandas==0.25.3
xlearn==0.4.4